Refactor TimerSystem code.
